Atualizando formato de menus e associando listas
